+++ Kernel/Output/Template/Provider.pm
@@ -475,6 +475,317 @@ sub _PreProcessTemplateContent {
 
 }
 
+=item MigrateDTLtoTT()
+
+translates old DTL template content to Template::Toolkit syntax.
+
+    my $TTCode = $ProviderObject->MigrateDTLtoTT( Content => $DTLCode );
+
+If an error was found, this method will die(), so please use eval around it.
+
+=cut
+
+sub MigrateDTLtoTT {
+    my ( $Self, %Param ) = @_;
+
+    my $Content = $Param{Content};
+
+    my $ID = "[a-zA-Z0-9:_\-]+";
+
+    my $SafeArrrayAccess = sub {
+        my $ID = shift;
+        if ( $ID !~ m{^[a-zA-Z0-9_]+$}xms ) {
+            return "item(\"$ID\")";
+        }
+        return $ID;
+    };
+
+    # $Quote $Config
+    $Content =~ s{\$Quote{"\$Config{"($ID)"}"}}{[% Config("$1") | html %]}smxg;
+
+    # $Quote $Env
+    $Content =~ s{\$Quote{"\$Env{"($ID)"}"}}{[% Env("$1") | html %]}smxg;
+
+    # $Quote $Data
+    $Content =~ s{
+            \$Quote{"\$Data{"($ID)"}"}
+        }
+        {
+            '[% Data.' . $SafeArrrayAccess->($1) . ' | html %]'
+        }esmxg;
+
+    # $Quote with length
+    $Content =~ s{
+            \$Quote{"\$Data{"($ID)"}",\s*"(\d+)"}
+        }
+        {
+            '[% Data.' . $SafeArrrayAccess->($1) . " | truncate($2) | html %]"
+        }esmxg;
+
+    # $Quote with dynamic length
+    $Content =~ s{
+            \$Quote{"\$Data{"($ID)"}",\s*"\$Q?Data{"($ID)"}"}
+        }
+        {
+            '[% Data.' . $SafeArrrayAccess->($1) . ' | truncate(Data.' . $SafeArrrayAccess->($2) . ') | html %]'
+        }esmxg;
+
+    # $Quote with translated text and fixed length
+    $Content =~ s{
+            \$Quote{"\$Text{"\$Data{"($ID)"}"}",\s*"(\d+)"}
+        }
+        {
+            '[% Data.' . $SafeArrrayAccess->($1) . " | Translate | truncate($2) | html %]"
+        }esmxg;
+
+    # $Quote with translated text and dynamic length
+    $Content =~ s{
+            \$Quote{"\$Text{"\$Data{"($ID)"}"}",\s*"\$Q?Data{"($ID)"}"}
+        }
+        {
+            '[% Data.' . $SafeArrrayAccess->($1) . ' | Translate | truncate(Data.' . $SafeArrrayAccess->($2) . ') | html %]'
+        }esmxg;
+
+    my $MigrateTextTag = sub {
+        my %Param       = @_;
+        my $Mode        = $Param{Mode};          # HTML or JSON
+        my $Text        = $Param{Text};          # The translated text
+        my $Dot         = $Param{Dot};           # Closing dot, sometimes outside of the Tag
+        my $ParamString = $Param{Parameters};    # Parameters to interpolate
+
+        my $Result = '[% ';
+
+        # Text contains a tag
+        if ( $Text =~ m{\$TimeLong{"\$Q?Data{"($ID)"}"}}smx ) {
+            $Result .= "Translate(Localize(Data." . $SafeArrrayAccess->($1) . ", \"TimeLong\")";
+        }
+        elsif ( $Text =~ m{\$TimeShort{"\$Q?Data{"($ID)"}"}}smx ) {
+            $Result .= "Translate(Localize(Data." . $SafeArrrayAccess->($1) . ", \"TimeShort\")";
+        }
+        elsif ( $Text =~ m{\$Date{"\$Q?Data{"($ID)"}"}}smx ) {
+            $Result .= "Translate(Localize(Data." . $SafeArrrayAccess->($1) . ", \"Date\")";
+        }
+        elsif ( $Text =~ m{\$Q?Data{"($ID)"}}smx ) {
+            $Result .= "Translate(Data." . $SafeArrrayAccess->($1) . "";
+        }
+        elsif ( $Text =~ m{\$Config{"($ID)"}}smx ) {
+            $Result .= "Translate(Config(\"$1\")";
+        }
+        elsif ( $Text =~ m{\$Q?Env{"($ID)"}}smx ) {
+            $Result .= "Translate(Env(\"$1\")";
+        }
+
+        # Plain text
+        else {
+            $Text =~ s{"}{\\"}smxg;    # Escape " signs
+            if ( $Param{Dot} ) {
+                $Text .= $Param{Dot};
+            }
+            $Result .= "Translate(\"$Text\"";
+        }
+
+        my @Parameters = split m{,\s*}, $ParamString;
+
+        PARAMETER:
+        for my $Parameter (@Parameters) {
+            next PARAMETER if ( !$Parameter );
+            if ( $Parameter =~ m{\$TimeLong{"\$Q?Data{"($ID)"}"}}smx ) {
+                $Result .= ", Localize(Data.$1, \"TimeLong\")";
+            }
+            elsif ( $Parameter =~ m{\$TimeShort{"\$Q?Data{"($ID)"}"}}smx ) {
+                $Result .= ", Localize(Data.$1, \"TimeShort\")";
+            }
+            elsif ( $Parameter =~ m{\$Date{"\$Q?Data{"($ID)"}"}}smx ) {
+                $Result .= ", Localize(Data.$1, \"Date\")";
+            }
+            elsif ( $Parameter =~ m{\$Q?Data{"($ID)"}}smx ) {
+                $Result .= ", Data.$1";
+            }
+            elsif ( $Parameter =~ m{\$Config{"($ID)"}}smx ) {
+                $Result .= ", Config(\"$1\")";
+            }
+            elsif ( $Parameter =~ m{\$Q?Env{"($ID)"}}smx ) {
+                $Result .= ", Env(\"$1\")";
+            }
+            else {
+                $Parameter =~ s{^"|"$}{}smxg;    # Remove enclosing ""
+                $Parameter =~ s{"}{\\"}smxg;     # Escape " signs in the string
+                $Result .= ", \"$Parameter\"";
+            }
+        }
+
+        if ( $Mode eq 'JSON' ) {
+            $Result .= ') | JSON %]';
+        }
+        else {
+            $Result .= ') | html %]';
+        }
+
+        return $Result;
+    };
+
+    my $TextOrData = "";
+
+    # $Text
+    $Content =~ s{
+            \$Text{
+                ["']
+                (
+                    [^\$]+?
+                    |\$Q?Data{\"$ID\"}
+                    |\$Config{\"$ID\"}
+                    |\$Q?Env{\"$ID\"}
+                    |\$TimeLong{\"\$Q?Data{\"$ID\"}\"}
+                    |\$TimeShort{\"\$Q?Data{\"$ID\"}\"}
+                    |\$Date{\"\$Q?Data{\"$ID\"}\"}
+                )
+                ["']
+                ((?:
+                    ,\s*["']
+                    (?:
+                        [^\$]+?
+                        |\$Q?Data{\"$ID\"}
+                        |\$Config{\"$ID\"}
+                        |\$Q?Env{\"$ID\"}
+                        |\$TimeLong{\"\$Q?Data{\"$ID\"}\"}
+                        |\$TimeShort{\"\$Q?Data{\"$ID\"}\"}
+                        |\$Date{\"\$Q?Data{\"$ID\"}\"}
+                    )
+                ["'])*)
+            }
+        }
+        {
+            $MigrateTextTag->( Mode => 'HTML', Text => $1, Parameters => $2);
+        }esmxg;
+
+    # drop empty $Text
+    $Content =~ s{\$Text{""}}{}xmsg;
+
+    # $JSText
+    $Content =~ s{
+            ["']\$JSText{
+                ["']
+                (
+                    [^\$]+?
+                    |\$Q?Data{\"$ID\"}
+                    |\$Config{\"$ID\"}
+                    |\$Q?Env{\"$ID\"}
+                    |\$TimeLong{\"\$Q?Data{\"$ID\"}\"}
+                    |\$TimeShort{\"\$Q?Data{\"$ID\"}\"}
+                    |\$Date{\"\$Q?Data{\"$ID\"}\"}
+                )
+                ["']
+                ((?:
+                    ,\s*["']
+                    (?:
+                        [^\$]+?
+                        |\$Q?Data{\"$ID\"}
+                        |\$Config{\"$ID\"}
+                        |\$Q?Env{\"$ID\"}
+                        |\$TimeLong{\"\$Q?Data{\"$ID\"}\"}
+                        |\$TimeShort{\"\$Q?Data{\"$ID\"}\"}
+                        |\$Date{\"\$Q?Data{\"$ID\"}\"}
+                    )
+                ["'])*)
+            }
+            (.?)["']
+        }
+        {
+            $MigrateTextTag->( Mode => 'JSON', Text => $1, Parameters => $2, Dot => $3);
+        }esmxg;
+
+    # $TimeLong
+    $Content =~ s{\$TimeLong{"\$Q?Data{"($ID)"}"}}{[% Data.$1 | Localize("TimeLong") %]}smxg;
+
+    # $TimeShort
+    $Content =~ s{\$TimeShort{"\$Q?Data{"($ID)"}"}}{[% Data.$1 | Localize("TimeShort") %]}smxg;
+
+    # $Date
+    $Content =~ s{\$Date{"\$Q?Data{"($ID)"}"}}{[% Data.$1 | Localize("Date") %]}smxg;
+
+    # $QData with length
+    $Content =~ s{
+            \$QData{"($ID)",\s*"(\d+)"}
+        }
+        {
+            "[% Data." . $SafeArrrayAccess->($1) . " | truncate($2) | html %]"
+        }esmxg;
+
+    # simple $QData
+    $Content =~ s{
+            \$QData{"($ID)"}
+        }
+        {
+            "[% Data." . $SafeArrrayAccess->($1) . " | html %]"
+        }esmxg;
+
+    # $LQData
+    $Content =~ s{
+            \$LQData{"($ID)"}
+        }
+        {
+            "[% Data." . $SafeArrrayAccess->($1) . " | uri %]"
+        }esmxg;
+
+    # simple $Data
+    $Content =~ s{
+            \$Data{"($ID)"}
+        }
+        {
+            "[% Data." . $SafeArrrayAccess->($1) . " %]"
+        }esmxg;
+
+    # $Config
+    $Content =~ s{\$Config{"($ID)"}}{[% Config("$1") %]}smxg;
+
+    # $Env
+    $Content =~ s{\$Env{"($ID)"}}{[% Env("$1") %]}smxg;
+
+    # $QEnv
+    $Content =~ s{\$QEnv{"($ID)"}}{[% Env("$1") | html %]}smxg;
+
+    # dtl:block
+    my %BlockSeen;
+    $Content =~ s{<!--\s*dtl:block:($ID)\s*-->}{
+        if ($BlockSeen{$1}++) {
+            "[% RenderBlockEnd(\"$1\") %]";
+        }
+        else {
+            "[% RenderBlockStart(\"$1\") %]";
+        }
+    }esmxg;
+
+    # dtl:js_on_document_complete
+    $Content =~ s{
+            <!--\s*dtl:js_on_document_complete\s*-->(.*?)<!--\s*dtl:js_on_document_complete\s*-->
+        }
+        {
+            "[% WRAPPER JSOnDocumentComplete %]${1}[% END %]";
+        }esmxg;
+
+    # dtl:js_on_document_complete_insert
+    $Content
+        =~ s{<!--\s*dtl:js_on_document_complete_placeholder\s*-->}{[% PROCESS JSOnDocumentCompleteInsert %]}smxg;
+
+    # $Include
+    $Content =~ s{\$Include{"($ID)"}}{[% InsertTemplate("$1.tt") %]}smxg;
+
+    my ( $Counter, $ErrorMessage );
+    LINE:
+    for my $Line ( split /\n/, $Content ) {
+        $Counter++;
+
+        # Make sure there are no more DTL tags present in the code.
+        if ( $Line =~ m{\$(?:L?Q?Data|Quote|Config|Q?Env|Time|Date|Text|JSText|Include)\{}xms ) {
+            $ErrorMessage .= "Line $Counter: $Line\n";
+        }
+    }
+
+    die $ErrorMessage if $ErrorMessage;
+
+    return $Content;
+}
+
 1;
 
 =back